move file index data to it’s own file
